backend/requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-CORS>=4.0.0
Flask-JWT-Extended>=4.5.0
matplotlib>=3.7.0
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.0
Werkzeug>=2.3.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
